Split file into two parts
